Add resolution and bezier input detection

+++ src/CurveGradient.hpp
@@ -15,10 +15,10 @@ class CurveGradient: public GradientAlgorithm {
     public:
     // calculates a gradient based on distance from a curve
     void applyGradient(image<rgba_pixel> &img) {
-        for (int x = 0; x < img.get_width(); x++) {
+        for (int x = 0; x < (int)img.get_width(); x++) {
             // apply a function
             int yVal = applyCos(x);
-            for (int y = 0; y < img.get_height(); y++) {
+            for (int y = 0; y < (int)img.get_height(); y++) {
                 // find distance from y to yPos, scale to color max, scale to image height
                 const double val = 255.0*abs(y - yVal)/img.get_height();
                 // apply blue gradient
