+++ src/HorizontalGradient.hpp
@@ -6,8 +6,8 @@
 class HorizontalGradient: public GradientAlgorithm {
     public:
     void applyGradient(image<rgba_pixel> &img) {
-        for (int x = 0; x < img.get_width(); x++) {
-            for (int y = 0; y < img.get_height(); y++) {
+        for (int x = 0; x < (int)img.get_width(); x++) {
+            for (int y = 0; y < (int)img.get_height(); y++) {
                 double r = 0;
                 double g = 255 * x/img.get_width();
                 double b = 255 - g;
